add serialize and deserialize test
